% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_documentation.R
\name{audit_documentation}
\alias{audit_documentation}
\title{Audit Roxygen Documentation Mismatches}
\usage{
audit_documentation(dir_path = ".")
}
\arguments{
\item{dir_path}{String. Project root.}
}
\value{
Dataframe of mismatches.
}
\description{
Checks if the function name in the \verb{@examples} block matches the actual function.
}
\author{
Mark London
}
