% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_exports.R
\name{audit_exports}
\alias{audit_exports}
\title{Audit Function Exports}
\usage{
audit_exports(inventory, dir_path = ".")
}
\arguments{
\item{inventory}{Dataframe. The output of audit_inventory().}

\item{dir_path}{String. Project root.}
}
\value{
The inventory dataframe with new columns: 'has_export_tag', 'in_namespace', 'status'.
}
\description{
Checks if functions are exported via Roxygen tags and if they appear in the NAMESPACE.
}
