% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_formula_functions.R
\name{audit_formula_functions}
\alias{audit_formula_functions}
\title{Audit Functions Used in Rules}
\usage{
audit_formula_functions(dir_path)
}
\arguments{
\item{dir_path}{String. Path to directory containing CSV rules.}
}
\value{
A frequency table of function calls (e.g., "mutate": 500, "if_else": 20).
}
\description{
Scans a directory of rule CSVs, parses the R code in them, and counts
which functions are actually being called.
}
