% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_package_clones.R
\name{audit_package_clones}
\alias{audit_package_clones}
\title{Generate Package Clone Report}
\usage{
audit_package_clones(pkg_path, output_file = "clone_report.xlsx")
}
\description{
Generate Package Clone Report
}
