% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smells_scoping.R
\name{check_free_variables}
\alias{check_free_variables}
\title{Check for Free Variables (Implicit Dependencies)}
\usage{
check_free_variables(pdata, file)
}
\arguments{
\item{pdata}{The AST dataframe.}

\item{file}{The filename.}
}
\description{
Finds variables used in a function that are neither Arguments nor Local Variables.
These are "Free Variables" that rely on Global State (or packages).
}
