% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rules.R
\name{compile_rules}
\alias{compile_rules}
\title{Compile Rules}
\usage{
compile_rules(file_path, allowed_vars = c("r", "tax", "net"))
}
\arguments{
\item{file_path}{String. Path to the CSV/Excel file.}

\item{allowed_vars}{Character Vector. The whitelist of allowed outputs.}
}
\value{
A 'recipe' data frame containing parsed expressions and dependencies.
}
\description{
Parses a configuration file into a 'Recipe' state machine.
Performs validation (typos, security) but DOES NOT execute.
}
\author{
Mark London
}
