% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_near_dupes.R
\name{detect_near_dupes}
\alias{detect_near_dupes}
\title{Detect Non-DRY Logic}
\usage{
detect_near_dupes(inv, threshold = 40, bucket_width = 100)
}
\arguments{
\item{inv}{Result from get_function_inventory}

\item{threshold}{Levenshtein distance limit (default 40)}
}
\description{
Detect Non-DRY Logic
}
