% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_safe_rules.R
\name{execute_safe_rules}
\alias{execute_safe_rules}
\title{Execute Safe Rules}
\usage{
execute_safe_rules(file_path, allowed_vars = c("r"))
}
\arguments{
\item{file_path}{String. Path to the CSV/Excel file.}

\item{allowed_vars}{Character Vector. The "Whitelist" of allowed outputs.}
}
\value{
A list containing the calculated results.
}
\description{
Reads a configuration file, validates variable names against a strict schema,
and executes the logic if safe.
}
\author{
Mark London
}
