% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_inventory.R
\name{get_function_inventory}
\alias{get_function_inventory}
\title{Extract Function DNA}
\usage{
get_function_inventory(target_dir)
}
\description{
Parses R files and returns a cleaned inventory of function bodies.
}
