% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_variables.R
\name{harvest_variables}
\alias{harvest_variables}
\title{Harvest Variables}
\usage{
harvest_variables(dir_path)
}
\arguments{
\item{dir_path}{String. Path to the directory containing rule files.}
}
\value{
A tibble with columns: variable, count, file_sources.
}
\description{
Scans a directory of configuration files (CSVs) and compiles a census
of all variables being assigned to (LHS of assignments).
}
\author{
Mark London
}
