% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_path_contract.R
\name{infer_path_contract}
\alias{infer_path_contract}
\title{Infer Path Contract}
\usage{
infer_path_contract(dir_path = ".", config_var = "paths")
}
\arguments{
\item{dir_path}{String. Project root.}

\item{config_var}{String. The name of your paths list variable (default "paths").}
}
\value{
A dataframe describing the contract (Input/Output) for each key.
}
\description{
Scans the code to determine which variables are treated as Inputs (Read)
and which are treated as Outputs (Write).
}
\author{
Mark London
}
