% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_recipe.R
\name{inspect_recipe}
\alias{inspect_recipe}
\title{Inspect Recipe for Missing Inputs}
\usage{
inspect_recipe(recipe)
}
\arguments{
\item{recipe}{A 'rule_recipe' object created by compile_rules().}
}
\value{
A list containing 'inputs_needed' and 'outputs_created'.
}
\description{
Static analysis of the recipe to determine which variables must be provided
by the user (Inputs) and which are calculated internally (Outputs).
}
\author{
Mark London
}
