% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_internal_calls.R
\name{map_internal_calls}
\alias{map_internal_calls}
\title{Map Internal Function Calls}
\usage{
map_internal_calls(inventory, dir_path = ".")
}
\arguments{
\item{inventory}{The dataframe from audit_inventory()}

\item{dir_path}{String. The project root (required to resolve relative paths in inventory).}
}
\value{
A dataframe of edges (from, to)
}
\description{
Scans the body of each function in the inventory to find calls to other project functions.
}
\author{
Mark London
}
