% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_audit_inventory.R
\name{new_audit_inventory}
\alias{new_audit_inventory}
\title{Constructor: Audit Inventory Object}
\usage{
new_audit_inventory(df)
}
\arguments{
\item{df}{Dataframe containing the raw inventory scan.}
}
\value{
An object of class 'audit_inventory'.
}
\description{
Validates and creates a standardized audit inventory object.
Enforces the schema to prevent silent column errors (The "Global Coupling" fix).
}
\author{
Mark London
}
