% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_smells.R
\name{print_smells}
\alias{print_smells}
\title{Print Code Smells Report}
\usage{
print_smells(smells)
}
\arguments{
\item{smells}{Dataframe. Output from detect_code_smells().}
}
\description{
Formats the output of detect_code_smells() into a readable CLI report.
}
\author{
Mark London
}
