% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_misplaced.R
\name{refactor_misplaced}
\alias{refactor_misplaced}
\title{Refactor Misplaced Functions (In Situ)}
\usage{
refactor_misplaced(dir_path = ".", dry_run = TRUE)
}
\arguments{
\item{dir_path}{String. Project root.}

\item{dry_run}{Logical. If TRUE, reports what would happen.}
}
\description{
Automatically moves functions into their own files.
}
