% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_functions.R
\name{relocate_functions}
\alias{relocate_functions}
\title{Relocate Functions to New Repositories}
\usage{
relocate_functions(inventory, function_map, source_root, dry_run = TRUE)
}
\arguments{
\item{inventory}{The dataframe returned by audit_inventory().}

\item{function_map}{A named list where keys are destination paths and values are character}

\item{source_root}{String. The root directory to read source files from.  <-- ADD THIS}

\item{dry_run}{Logical. If TRUE, just logs actions.}
}
\description{
Reads the project inventory and copies specific functions into a new directory structure.
Useful for splitting a Monolith into Micro-services.
}
