% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recipe.R
\name{run_recipe}
\alias{run_recipe}
\title{Run Recipe}
\usage{
run_recipe(recipe)
}
\arguments{
\item{recipe}{A 'rule_recipe' object created by compile_rules().}
}
\value{
A list containing the calculated results.
}
\description{
Executes a compiled 'rule_recipe' object in a safe environment.
}
\author{
Mark London
}
