% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_io_logic.R
\name{scan_io_logic}
\alias{scan_io_logic}
\title{Scan I/O Logic (Reads, Writes, and Path Definitions)}
\usage{
scan_io_logic(dir_path = ".")
}
\arguments{
\item{dir_path}{String. Path to project root.}
}
\value{
A dataframe of I/O events.
}
\description{
Parses code to find data ingestion, data output, and path construction.
}
\author{
Mark London
}
