% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_flow.R
\name{visualize_flow}
\alias{visualize_flow}
\title{Visualize Downstream Calls (Forward Dependency)}
\usage{
visualize_flow(target_func, dir_path = ".", save_dot = NULL)
}
\arguments{
\item{target_func}{String. The entry point function.}

\item{dir_path}{String. Project root.}

\item{save_dot}{String. Optional path to save the DOT file.}
}
\value{
A DiagrammeR graph object.
}
\description{
Generates a flow chart starting from a top-level function to see
everything it calls, and what they call in turn.
}
