% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_performance.R
\name{visualize_performance}
\alias{visualize_performance}
\title{Visualize Performance Heatmap}
\usage{
visualize_performance(target_func, dynamic_data, dir_path = ".")
}
\arguments{
\item{target_func}{String. The entry point function to trace (e.g. "nth_prime_bad").}

\item{dynamic_data}{Dataframe. The output from rdyntrace::trace_results().}

\item{dir_path}{String. Project root.}
}
\value{
A DiagrammeR graph object.
}
\description{
Generates a call graph where nodes are colored by their execution time.
Red = Slow, Green = Fast, Grey = Unused.
}
\author{
Mark London
}
