% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_rules.R
\name{visualize_rules}
\alias{visualize_rules}
\title{Visualize Rules from Recipe}
\usage{
visualize_rules(recipe)
}
\arguments{
\item{recipe}{A 'rule_recipe' object created by compile_rules().}
}
\value{
A DiagrammeR graph object.
}
\description{
Generates a flowchart from a compiled Recipe object.
}
\author{
Mark London
}
