% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_inventory.R
\name{audit_inventory}
\alias{audit_inventory}
\title{Audit Project Inventory}
\usage{
audit_inventory(dir_path)
}
\arguments{
\item{dir_path}{String. Path to the package root.}
}
\value{
A tibble listing objects, their locations, coverage, and placement status.
}
\description{
Scans a project to index all defined functions and checks test coverage.
Also validates if functions are in correctly named files.
}
