% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_rules.R
\name{compile_rules}
\alias{compile_rules}
\title{Compile Rules}
\usage{
compile_rules(file_path, allowed_vars = NULL)
}
\arguments{
\item{file_path}{String. Path to the CSV/Excel file.}

\item{allowed_vars}{Character Vector. The whitelist of allowed outputs.
If NULL (default), strict checking is disabled (Discovery Mode).}
}
\description{
Parses a configuration file into a 'Recipe' state machine.
Performs validation (typos, security) but DOES NOT execute.
}
\author{
Mark London
}
