% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_progress.R
\name{visualize_progress}
\alias{visualize_progress}
\title{Visualize Migration Progress}
\usage{
visualize_progress(inventory)
}
\arguments{
\item{inventory}{A dataframe from audit_inventory().}
}
\description{
Draws a network graph showing the decoupling of the legacy monolith.
}
