% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_rules.R
\name{visualize_rules}
\alias{visualize_rules}
\title{Visualize Rules}
\usage{
visualize_rules(file_path)
}
\arguments{
\item{file_path}{String. Path to the CSV/Excel file.}
}
\value{
A DiagrammeR graph object.
}
\description{
Reads a configuration file and generates a visual flowchart of the logic.
}
\author{
Mark London
}
